Create new tests folder
