+++ src/libs/llm_utils.py
@@ -81,19 +81,16 @@ Now classify this email:
                         "type": "object",
                         "properties": {
                             "category": {
-                                "type": "array",
-                                "items": {
-                                    "type": "string",
-                                    "enum": [
-                                        "urgent",
-                                        "actionable",
-                                        "information",
-                                        "newsletter",
-                                        "promo",
-                                        "other",
-                                    ],
-                                },
-                                "description": "One or more categories",
+                                "type": "string",
+                                "enum": [
+                                    "urgent",
+                                    "actionable",
+                                    "information",
+                                    "newsletter",
+                                    "promo",
+                                    "other",
+                                ],
+                                "description": "One category",
                             },
                         },
                         "required": ["category"],
@@ -109,45 +106,3 @@ Now classify this email:
 
     except Exception as e:
         return ["error"], str(e)
-
-def test_llm_classifications(classifier_fn, data_path="src/libs/email_class_data.json"):
-    """
-    classifier_fn: function(email_text) -> list of predicted categories (strings)
-    data_path: path to your labeled JSON data file
-    
-    Prints accuracy and mismatches.
-    """
-    with open(data_path, "r", encoding="utf-8") as f:
-        data = json.load(f)
-    
-    total = len(data)
-    correct = 0
-    mismatches = []
-    
-    for i, entry in enumerate(data):
-        text = entry["text"]
-        true_classes = entry.get("categories", [])
-        
-        pred_classes, _ = classifier_fn(text)
-        
-        if pred_classes == true_classes:
-            correct += 1
-        else:
-            mismatches.append({
-                "index": i,
-                "true": true_classes,
-                "predicted": pred_classes,
-                "text_snippet": text[:200].replace("\n", " ") + "..."
-            })
-    
-    accuracy = correct / total if total > 0 else 0
-    print(f"Total samples: {total}")
-    print(f"Correct classifications: {correct}")
-    print(f"Accuracy: {accuracy:.2%}")
-    
-    if mismatches:
-        print(f"\nMismatches ({len(mismatches)}):")
-        for m in mismatches:
-            print(f"- Index {m['index']}: True={m['true']}, Predicted={m['predicted']}")
-            print(f"  Text snippet: {m['text_snippet']}")
-            print()
